// front/src/styles/components/Modals/ModalAddSpot.module.scss.d.ts
export type Styles = {
  add_spot_botton: string
  button_round: string
  button_wrapper: string
  message: string
  modal_content: string
  modal_overlay: string
  ReactModal__Overlay: string
  spot_address: string
  spot_data: string
  spot_name: string
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles